+++ example/provider.dart
@@ -8,7 +8,7 @@ void main() {
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return Provider(
+    return Provider.value(
       value: 42,
       child: Example(),
     );