+++ lib/provider.dart
@@ -2,4 +2,4 @@ library provider;
 
 export 'src/consumer.dart';
 export 'src/listenable_provider.dart';
-export 'src/provider.dart' hide debugGetProviderUpdateShouldNotify;
+export 'src/provider.dart';